chore(rename): 统一命名修复拼写错误
- LiteSoritePlugin → LiteSortiePlugin (Sorite→Sortie拼写)
- NighWavePlugin → NightWavePlugin (Nigh→Night拼写)
- MarketLichsSisterUtils → MarketLichSisterUtils (Lichs→Lich统一)
